validatie

+++ src/main/java/be/odisee/team5/fsfopdracht2/controllers/BestellingController.java
@@ -6,11 +6,14 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+
+import javax.validation.Valid;
 
 @Slf4j
 @Controller
@@ -42,7 +45,7 @@ public class BestellingController {
      * @param bestellingData the data for the entry to be saved
      */
     @PostMapping(params = "submit")
-    public String processBestelling(BestellingData bestellingData, Errors errors, Model model) {
+    public String processBestelling( @Valid  BestellingData bestellingData, BindingResult errors, Model model) {
 
         String message="";
 
@@ -71,9 +74,12 @@ public class BestellingController {
      * When delete of an entry is pressed
      */
     @PostMapping(params = "delete")
-    public String deleteBestelling(BestellingData bestellingData, Model model){
+    public String deleteBestelling(@Valid  BestellingData bestellingData, Model model, BindingResult result){
+
+        if(result.hasErrors()) return "/delete";
+
         fsfService.deleteBestelling(bestellingData.getId());
-        BestellingData bD = fsfService.prepareNewBestellingData();
+        //BestellingData bD = fsfService.prepareNewBestellingData();
         bestellingData = fsfService.prepareNewBestellingData();
         prepareForm(bestellingData, model);
         return "createBestelling";
@@ -89,10 +95,14 @@ public class BestellingController {
     @GetMapping("/edit")
     public String bestellingEditForm(@RequestParam("id") long id, Model model) {
 
-        BestellingData bestellingData = fsfService.prepareEntryDataToEdit(id);
-        prepareForm(bestellingData, model);
-        model.addAttribute("message", "Update or Delete this entry please - or Cancel");
-        return "createBestelling";
+
+
+
+            BestellingData bestellingData = fsfService.prepareEntryDataToEdit(id);
+            prepareForm(bestellingData, model);
+            model.addAttribute("message", "Update or Delete this entry please - or Cancel");
+            return "createBestelling";
+
     }
 
     /**
@@ -102,5 +112,28 @@ public class BestellingController {
     @PostMapping(params = "cancel")
     public String redirectToCreate() {
         return "redirect:/bestelling";
+    }
+
+    @PostMapping(params = "inplannen")
+    public String inplannenProcess(@Valid  BestellingData bestellingData, BindingResult errors, Model model){
+        String message="";
+        try {
+            // Are there any input validation errors detected by JSR 380 bean validation?
+            if (errors.hasErrors() ) {
+                message = "Correct input errors, please";
+                throw new IllegalArgumentException();
+            }
+            // Now that the input seems to be OK, let's create a new entry or update/delete an existing entry
+            message = fsfService.processBestellingInplannen(bestellingData);
+
+            // Prepare form for new data-entry
+            bestellingData = fsfService.prepareNewBestellingData();
+
+        } catch (IllegalArgumentException e) {
+            // Nothing special needs to be done
+        }
+        prepareForm(bestellingData, model);
+        model.addAttribute("message", message);
+        return "createBestelling";
     }
 }
